Extracts duplicate code to helper files in shared scope.
