+++ FeedApplication/FeedApplicationTests/ImageLoaderCacheDecoratorTests.swift
@@ -20,7 +20,7 @@ class ImageLoaderCacheDecorator: ImageLoader {
     }
 }
 
-class ImageLoaderCacheDecoratorTests: XCTestCase {
+class ImageLoaderCacheDecoratorTests: XCTestCase, ImageLoaderTestCase {
     
     func test_init_doesNotLoadImageData() {
         let (_, loader) = makeSUT()
@@ -74,51 +74,7 @@ class ImageLoaderCacheDecoratorTests: XCTestCase {
         return (sut, loader)
     }
     
-    private func expect(_ sut: ImageLoader, toCompleteWith expectedResult: ImageLoader.Result, when action: () -> Void, file: StaticString = #file, line: UInt = #line) {
-        let exp = expectation(description: "Wait for load completion")
-        
-        _ = sut.loadImageData(from: anyURL()) { receivedResult in
-            switch (receivedResult, expectedResult) {
-            case let (.success(receivedImageData), .success(expectedImageData)):
-                XCTAssertEqual(receivedImageData, expectedImageData, file: file, line: line)
-            case (.failure, .failure):
-                break
-            default:
-                 XCTFail("Expected \(expectedResult), got \(receivedResult) instead", file: file, line: line)
-            }
-            exp.fulfill()
-        }
-        action()
-        wait(for: [exp], timeout: 3.0)
-    }
-    
     private func anyData() -> Data {
         return Data("data".utf8)
-    }
-    
-    private class ImageLoaderSpy: ImageLoader {
-        private struct Task: ImageLoaderTask {
-            let callback: () -> Void
-            func cancel() { callback() }
-        }
-        
-        private var loadRequests = [(url: URL, completion: (ImageLoader.Result) -> Void )]()
-        var loadedURLs: [URL] {  return loadRequests.map { $0.url } }
-        private(set) var cancelledURLs = [URL]()
-        
-        func loadImageData(from url: URL, completion: @escaping (ImageLoader.Result) -> Void) -> ImageLoaderTask {
-            loadRequests.append((url, completion))
-            return Task { [weak self] in
-                self?.cancelledURLs.append(url)
-            }
-        }
-        
-        func complete(with error: Error, at index: Int = 0) {
-            loadRequests[index].completion(.failure(error))
-        }
-        
-        func complete(with data: Data, at index: Int = 0) {
-            loadRequests[index].completion(.success(data))
-        }
     }
 }
